requirements.txt
manga-ocr>=0.1.11
gradio>=4.0.0
Pillow>=10.0.0
